refactor: move Employment component hooks and types to hooks.ts file
